Backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.1
sqlalchemy==2.0.36
aiosqlite==0.22.0
alembic==1.13.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.1
